ran forge fmt

+++ test/utils/DiamondManager.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: Unlicense
 pragma solidity ^0.8.0;
 
-import {Script} from "forge-std/Script.sol";
-import {IDiamondCut} from "src/diamond/IDiamondCut.sol";
-import {Diamond} from "src/diamond/Diamond.sol";
-import {DiamondCutFacet} from "src/diamond/DiamondCutFacet.sol";
-import {DiamondLoupeFacet} from "src/diamond/DiamondLoupeFacet.sol";
-import {AccessControlFacet} from "src/access/AccessControlFacet.sol";
-import {OwnershipFacet} from "src/access/OwnershipFacet.sol";
-import {PausableFacet} from "src/security/PausableFacet.sol";
-import {MetaTxFacet} from "src/metatx/MetaTxFacet.sol";
+import { Script } from "forge-std/Script.sol";
+import { IDiamondCut } from "src/diamond/IDiamondCut.sol";
+import { Diamond } from "src/diamond/Diamond.sol";
+import { DiamondCutFacet } from "src/diamond/DiamondCutFacet.sol";
+import { DiamondLoupeFacet } from "src/diamond/DiamondLoupeFacet.sol";
+import { AccessControlFacet } from "src/access/AccessControlFacet.sol";
+import { OwnershipFacet } from "src/access/OwnershipFacet.sol";
+import { PausableFacet } from "src/security/PausableFacet.sol";
+import { MetaTxFacet } from "src/metatx/MetaTxFacet.sol";
 
 struct FacetInfo {
     address addr;
@@ -50,18 +50,25 @@ contract DiamondManager is Script {
         return createDiamondAndInit(_facets, new Diamond.Initialization[](0));
     }
 
-    function createDiamondAndInit(FacetInfo[] memory _facets, Diamond.Initialization[] memory _inits) internal returns (Diamond) {
+    function createDiamondAndInit(
+        FacetInfo[] memory _facets,
+        Diamond.Initialization[] memory _inits
+    ) internal returns (Diamond) {
         return createDiamondAndInit(_facets, _inits, new FacetInfo[](0));
     }
 
-    function createDiamondAndInit(FacetInfo[] memory _facets, Diamond.Initialization[] memory _inits, FacetInfo[] memory _optionalFacets) internal returns (Diamond) {
+    function createDiamondAndInit(
+        FacetInfo[] memory _facets,
+        Diamond.Initialization[] memory _inits,
+        FacetInfo[] memory _optionalFacets
+    ) internal returns (Diamond) {
         DiamondCutFacet diamondCut = new DiamondCutFacet();
         DiamondLoupeFacet diamondLoupe = new DiamondLoupeFacet();
         OwnershipFacet ownership = new OwnershipFacet();
         AccessControlFacet accessControl = new AccessControlFacet();
         PausableFacet pausable = new PausableFacet();
         MetaTxFacet meta = new MetaTxFacet();
-        
+
         FacetInfo[6] memory staticFacets = [
             FacetInfo(address(diamondCut), "DiamondCutFacet", IDiamondCut.FacetCutAction.Add),
             FacetInfo(address(diamondLoupe), "DiamondLoupeFacet", IDiamondCut.FacetCutAction.Add),
@@ -71,32 +78,26 @@ contract DiamondManager is Script {
             FacetInfo(address(meta), "MetaTxFacet", IDiamondCut.FacetCutAction.Add)
         ];
 
-        IDiamondCut.FacetCut[] memory cuts = new IDiamondCut.FacetCut[](staticFacets.length + _facets.length + _optionalFacets.length);
+        IDiamondCut.FacetCut[] memory cuts =
+            new IDiamondCut.FacetCut[](staticFacets.length + _facets.length + _optionalFacets.length);
 
-        for (uint i = 0; i < staticFacets.length; i++) {
+        for (uint256 i = 0; i < staticFacets.length; i++) {
             cuts[i] = IDiamondCut.FacetCut(
-                staticFacets[i].addr,
-                staticFacets[i].action,
-                generateSelectors(staticFacets[i].name)
+                staticFacets[i].addr, staticFacets[i].action, generateSelectors(staticFacets[i].name)
             );
         }
 
-        for (uint i = 0; i < _facets.length; i++) {
-            if(_facets[i].addr == address(0)) {
+        for (uint256 i = 0; i < _facets.length; i++) {
+            if (_facets[i].addr == address(0)) {
                 revert("TEST ARGS INVALID: Invalid FacetInfo given");
             }
-            cuts[i + staticFacets.length] = IDiamondCut.FacetCut(
-                _facets[i].addr,
-                _facets[i].action,
-                generateSelectors(_facets[i].name)
-            );
+            cuts[i + staticFacets.length] =
+                IDiamondCut.FacetCut(_facets[i].addr, _facets[i].action, generateSelectors(_facets[i].name));
         }
 
-        for (uint i = 0; i < _optionalFacets.length; i++) {
-            cuts[i  + staticFacets.length + _facets.length] = IDiamondCut.FacetCut(
-                _optionalFacets[i].addr,
-                _optionalFacets[i].action,
-                generateSelectors(_optionalFacets[i].name)
+        for (uint256 i = 0; i < _optionalFacets.length; i++) {
+            cuts[i + staticFacets.length + _facets.length] = IDiamondCut.FacetCut(
+                _optionalFacets[i].addr, _optionalFacets[i].action, generateSelectors(_optionalFacets[i].name)
             );
         }
 
@@ -106,8 +107,8 @@ contract DiamondManager is Script {
             initData: abi.encodeWithSelector(AccessControlFacet.AccessControlFacet_init.selector)
         });
 
-        for (uint i = 0; i < _inits.length; i++) {
-            _initsAll[i+1] = _inits[i];
+        for (uint256 i = 0; i < _inits.length; i++) {
+            _initsAll[i + 1] = _inits[i];
         }
 
         return new Diamond(address(this), cuts, _initsAll);
